AND-2502 Move enroll biometrics dialog to save wallet feature package
