+++ app/src/main/java/com/tangem/tap/features/saveWallet/ui/SaveWalletScreenState.kt
@@ -1,8 +1,8 @@
 package com.tangem.tap.features.saveWallet.ui
 
 import androidx.compose.runtime.Immutable
-import com.tangem.core.ui.models.EnrollBiometricsDialog
 import com.tangem.tap.features.details.ui.cardsettings.TextReference
+import com.tangem.tap.features.saveWallet.ui.models.EnrollBiometricsDialog
 
 @Immutable
 internal data class SaveWalletScreenState(
